# attendance-system/backup_before_restore_20260716_200909/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Attendance System API
  description: >
    API quản lý chấm công tập trung, tích hợp nhiều dòng máy (ZKTeco, Sunbeam/Timmy, Hikvision).
    Hệ thống chỉ lưu trữ raw log chấm công và đồng bộ nhân viên hai chiều — không tính công.
  version: "1.0.0"
servers:
  - url: /api/v1

# JWT Bearer Auth áp dụng cho toàn bộ API (trừ /auth/login được override bên dưới)
security:
  - BearerAuth: []
...
paths:
...
  /shifts:
    post:
      tags: [shifts]
      summary: Tạo ca làm việc mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShiftRequest'
      responses:
        '201':
          description: Ca làm việc đã được tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
    get:
      tags: [shifts]
      summary: Danh sách ca làm việc
      responses:
        '200':
          description: Danh sách ca làm việc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
  /shifts/{id}:
    delete:
      tags: [shifts]
      summary: Xoá ca làm việc
      parameters:
        - $ref: '#/components/parameters/ShiftID'
      responses:
        '204':
          description: Đã xoá ca làm việc
  /employees/{id}/shifts:
    post:
      tags: [shifts]
      summary: Gán ca làm việc cho nhân viên
      parameters:
        - $ref: '#/components/parameters/EmployeeID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shift_id, start_date]
              properties:
                shift_id: { type: string }
                start_date: { type: string, format: date, example: "2026-07-01" }
                end_date: { type: string, format: date, example: "2026-12-31" }
      responses:
        '201':
          description: Ca đã được gán thành công

  # ─── Xin nghỉ phép (Leave Requests) ─────────────────────────────────────────
  /leave-requests:
    post:
      tags: [leave]
      summary: Đăng ký nghỉ phép
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeaveRequest'
      responses:
        '201':
          description: Đã tạo yêu cầu nghỉ phép
    get:
      tags: [leave]
      summary: Danh sách đơn xin nghỉ phép
      parameters:
        - in: query
          name: employee_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        '200':
          description: Danh sách đơn xin nghỉ phép
  /leave-requests/{id}/approve:
    post:
      tags: [leave]
      summary: Duyệt đơn nghỉ phép
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Đã duyệt
  /leave-requests/{id}/reject:
    post:
      tags: [leave]
      summary: Từ chối đơn nghỉ phép
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Đã từ chối

  # ─── Làm thêm giờ (Overtime Requests) ───────────────────────────────────────
  /overtime-requests:
    post:
      tags: [overtime]
      summary: Đăng ký làm thêm giờ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOvertimeRequest'
      responses:
        '201':
          description: Đã tạo yêu cầu OT
    get:
      tags: [overtime]
      summary: Danh sách đơn OT
      parameters:
        - in: query
          name: employee_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        '200':
          description: Danh sách đơn OT
  /overtime-requests/{id}/approve:
    post:
      tags: [overtime]
      summary: Duyệt đơn OT
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Đã duyệt
  /overtime-requests/{id}/reject:
    post:
      tags: [overtime]
      summary: Từ chối đơn OT
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Đã từ chối

  # ─── Bảng công tổng hợp (Daily Attendance) ───────────────────────────────────
  /daily-attendance/process:
    post:
      tags: [attendance]
      summary: Tính công cho tất cả nhân viên trong ngày chỉ định
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                  example: "2026-07-14"
      responses:
        '200':
          description: Tính công thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed:
                    type: integer
                    description: Số nhân viên đã tính công
  /daily-attendance/report:
    get:
      tags: [attendance]
      summary: Báo cáo bảng công tổng hợp (lọc theo nhân viên, khoảng thời gian)
      parameters:
        - in: query
          name: employee_id
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Bảng công tổng hợp
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyAttendance'

  # ─── Audit Log ───────────────────────────────────────────────────────────────
  /audit-logs:
    get:
      tags: [audit]
      summary: Danh sách nhật ký hoạt động hệ thống
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Danh sách audit log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLog'

  # ─── Dashboard & Real-time ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Lấy token từ POST /auth/login.
        Gửi kèm header: Authorization: Bearer <token>
  parameters:
...
    EmployeeID:
      name: id
      in: path
      required: true
      schema: { type: string }
    ShiftID:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    CreateShiftRequest:
      type: object
      required: [name, start_time, end_time]
      properties:
        name: { type: string, example: "Ca sáng" }
        start_time: { type: string, example: "08:00" }
        end_time: { type: string, example: "17:00" }
        break_minutes: { type: integer, default: 60 }
        late_grace_minutes: { type: integer, default: 5 }
        early_grace_minutes: { type: integer, default: 5 }
        max_working_minutes: { type: integer, default: 480 }
        timezone: { type: string, default: "Asia/Ho_Chi_Minh" }
    Shift:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        start_time: { type: string }
        end_time: { type: string }
        break_minutes: { type: integer }
        late_grace_minutes: { type: integer }
        early_grace_minutes: { type: integer }
        max_working_minutes: { type: integer }
        timezone: { type: string }
        created_at: { type: string, format: date-time }
    CreateLeaveRequest:
      type: object
      required: [employee_id, leave_type, start_date, end_date]
      properties:
        employee_id: { type: string }
        leave_type: { type: string, enum: [annual, sick, unpaid, business_trip] }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        reason: { type: string }
    CreateOvertimeRequest:
      type: object
      required: [employee_id, date, start_time, end_time]
      properties:
        employee_id: { type: string }
        date: { type: string, format: date }
        start_time: { type: string, example: "17:00" }
        end_time: { type: string, example: "20:00" }
    DailyAttendance:
      type: object
      properties:
        id: { type: integer }
        employee_id: { type: string }
        date: { type: string, format: date }
        shift_id: { type: string }
        first_in: { type: string, format: date-time }
        last_out: { type: string, format: date-time }
        late_minutes: { type: integer }
        early_minutes: { type: integer }
        working_hours: { type: number }
        attendance_status:
          type: string
          enum: [present, late, early, absent, leave]
        overtime_minutes: { type: integer }
        leave_id: { type: string }
    AuditLog:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: string }
        action: { type: string }
        object_type: { type: string }
        object_id: { type: string }
        description: { type: string }
        ip_address: { type: string }
        created_at: { type: string, format: date-time }
...